+++ dataworkspace/dataworkspace/apps/applications/urls_visualisations.py
@@ -6,6 +6,7 @@ from dataworkspace.apps.applications.views import (
     visualisation_branch_html_view,
     visualisation_users_give_access_html_view,
     visualisation_users_with_access_html_view,
+    visualisation_catalogue_item_html_view,
 )
 
 urlpatterns = [
@@ -24,5 +25,10 @@ urlpatterns = [
         '<str:gitlab_project_id>/users/with-access',
         login_required(visualisation_users_with_access_html_view),
         name='users-with-access',
+    ),
+    path(
+        '<str:gitlab_project_id>/catalogue-item',
+        login_required(visualisation_catalogue_item_html_view),
+        name='catalogue-item',
     ),
 ]